% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateResults}
\alias{annotateResults}
\title{annotate DESeq2 results}
\usage{
annotateResults(results, annotationFile)
}
\arguments{
\item{results}{DESeq2 results data.frame, from \code{\link{DESeq}}, \code{\link{results}}}

\item{annotationFile}{sliding window annotation file, can be plain either text or .gz file,
the file MUST be TAB separated, and MUST have the following columns:\cr
\code{chromosome} chromosome name \cr
\code{unique_id} unique id of the window \cr
\code{begin} window start co-ordinate \cr
\code{end} window end co-ordinate \cr
\code{strand} strand \cr
\code{gene_id} gene id \cr
\code{gene_name} gene name \cr
\code{gene_type} gene type annotation \cr
\code{gene_region} gene region \cr
\code{Nr_of_region} number of the current region \cr
\code{Total_nr_of_region} total number of regions \cr
\code{window_number} window number \cr}
}
\value{
data.frame
}
\description{
annotate DESeq2 results using annotation file from htseq-clip
}
