% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{countsPerRegion}
\alias{countsPerRegion}
\title{normalized counts per region}
\usage{
countsPerRegion(windowRes, padjCol = "padj", padjThresh = 0.05,
  log2FoldChangeCol = "log2FoldChange", log2FoldChangeThresh = 1,
  begin0based = TRUE, normalizedCounts, treatmentCols,
  treatmentName = "treatment", controlName = "control", op = "max")
}
\description{
given window resutls and normalized counts, combine significant overlapping windows into regions and
for each region, pick two candidate winodws:\cr (i) with highest log2FoldChange and\cr (ii) with highest normalized mean in
treatment samples.\cr Return a data.frame with region information and stats, and for the selected windows stats:\cr
normalized mean expression in treatment and control samples and\cr individual expression in replicates
}
