% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegions.R
\name{createRegions}
\alias{createRegions}
\title{create regions from significant windows}
\usage{
createRegions(mergeDat, padjCol = "pBonferroni.adj", padjThresh = 0.05,
  log2FoldChangeCol = "log2FoldChange", log2FoldChangeThresh = 1,
  ncores = 5)
}
\arguments{
\item{mergeDat}{data.frame, output from \code{\link{mergeWindows}}}

\item{padjCol}{name of the adjusted pvalue column (default: pBonferroni.adj)}

\item{padjThresh}{threshold for p-adjusted value (default: 0.05)}

\item{log2FoldChangeCol}{name of the log2foldchange column (default: log2FoldChange)}

\item{log2FoldChangeThresh}{threshold for log2foldchange value (default:1)}

\item{ncores}{number of cores to uses}
}
\value{
data.frame
}
\description{
create significant regions by merging significant windows
given p-adjusted value and log2 fold change columns and thresholds
}
