% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regions.R
\name{createRegionsHelper}
\alias{createRegionsHelper}
\title{: helper function, create region}
\usage{
createRegionsHelper(mergeDat, padjCol, padjThresh = 0.05,
  log2FoldChangeCol = "log2FoldChange", log2FoldChangeThresh = 1)
}
\arguments{
\item{mergeDat}{a gene specific result data.frame}

\item{padjCol}{name of the adjusted pvalue column}

\item{padjThresh}{threshold for p-adjusted value}

\item{log2FoldChangeCol}{name of the log2foldchange column}

\item{log2FoldChangeThresh}{threshold for log2foldchange value}

\item{ncores}{number of cores to uses}
}
\description{
: merge significant neighboring windows into regions
}
\keyword{internal}
