% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{DESeqDataSetFromSlidingWindows}
\alias{DESeqDataSetFromSlidingWindows}
\title{create DESeq data object}
\usage{
DESeqDataSetFromSlidingWindows(countData, colData, design, tidy = FALSE,
  ignoreRank = FALSE, annotObj, start0based = TRUE)
}
\arguments{
\item{countData}{sliding window count data}

\item{colData}{phenotype data}

\item{design}{design of the experiment, see \code{\link[DESeq2:DESeqDataSet]{DESeqDataSet}}}

\item{tidy}{If TRUE, first column is of countData is treated as rownames (defalt: FALSE), see \code{\link[DESeq2:DESeqDataSet]{DESeqDataSet}}}

\item{ignoreRank}{ignore rank, see \code{\link[DESeq2:DESeqDataSet]{DESeqDataSet}}}

\item{annotObj}{can either be a data.frame or a file name, see details}

\item{start0based}{TRUE (default) or FALSE. If TRUE, then the start positions in \code{annotObj} is  considered to be 0-based}
}
\value{
DESeq object
}
\description{
create DESeq data object from sliding window counts,
phenotype data and annotation data
}
\details{
If \code{annotObj} is a file name, the input file MUST be <TAB> separated, and supports reading in .gz files. If \code{annotObj} is a data.frame,
\code{colnames(annotObj)} MUST not be empty.\cr
This function checks for the following columns after reading in the file or on data.frame:\cr
 chromosome: chromosome name \cr
 unique_id: unique id of the window \cr
 begin: window start co-ordinate \cr
 end: window end co-ordinate \cr
 strand: strand \cr
 gene_id: gene id \cr
 gene_name: gene name \cr
 gene_type: gene type annotation \cr
 gene_region: gene region \cr
 Nr_of_region: number of the current region \cr
 Total_nr_of_region: total number of regions \cr
 window_number: window number \cr
}
