% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{.readAnnotation}
\alias{.readAnnotation}
\title{read annotation data}
\usage{
.readAnnotation(fname, uniqIds = NULL)
}
\arguments{
\item{fname}{file name/path}

\item{uniqIds}{filter stable and keep annotation for these unique ids}
}
\description{
read annotation data for windows
The file MUST be tab separated and MUST have the following columns:\cr
chromosome: chromosome name \cr
unique_id: unique id of the window \cr
begin: window start co-ordinate \cr
end: window end co-ordinate \cr
strand: strand \cr
gene_id: gene id \cr
gene_name: gene name \cr
gene_type: gene type annotation \cr
gene_region: gene region \cr
Nr_of_region: number of the current region \cr
Total_nr_of_region: total number of regions \cr
window_number: window number \cr
}
