% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_merger.R, R/window_merger2.R
\name{mergeWindows}
\alias{mergeWindows}
\title{merge windows}
\usage{
mergeWindows(annRes, minOverlap = 5, padjMethod = "BH")

mergeWindows(annRes, minOverlap = 5, padjMethod = "BH")
}
\arguments{
\item{annRes}{annotated DESEq2 table}

\item{minOverlap}{minimum overlap between sliding windows}

\item{padjMethod}{p-value adjustment method (multiple testing correction) for the whole table, can be any input for base function p.adjust}

\item{minDist}{minimum distance between the sliding windows
default 0 to account for bed formatted window regions}

\item{padjWindow}{p-value adjustment method for window, can be one of 'local' or 'bonferroni'}

\item{ncores}{number of cores to use}

\item{annRes}{annotated DESEq2 table}
}
\description{
merge annotated sliding windows

merge annotated sliding windows
}
