% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_merger.R
\name{mergeWindowsHelper_local}
\alias{mergeWindowsHelper_local}
\title{local simes correction}
\usage{
mergeWindowsHelper_local(annRes, minDist = 0)
}
\arguments{
\item{annRes}{annotated results (data.frame)}

\item{minDist}{minimum distance between non overlapping windows,
default: 0 to account for bed formatted window regions}
}
\description{
Helper function, merge overlapping windows and
correct p-values for these overlapping windows using local simes like method
}
\keyword{internal}
