% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_merger.R
\name{pvalue_dependence_correction}
\alias{pvalue_dependence_correction}
\title{pvalue correction using local simes like approach}
\usage{
pvalue_dependence_correction(p_to_correct, dependent_p_values)
}
\arguments{
\item{p_to_correct}{pvalue to correct (from the current window)}

\item{dependent_p_values}{pvalues from neighboring windows}
}
\description{
pvalue correction using local simes like approach
}
\keyword{internal}
