% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsDEWSeq.R
\name{resultsDEWSeq}
\alias{resultsDEWSeq}
\title{extract DEWseq results}
\usage{
resultsDEWSeq(object, annotationFile, contrast, name, listValues = c(1,
  -1), cooksCutoff, test, addMLE = FALSE, tidy = FALSE,
  parallel = FALSE, BPPARAM = bpparam(), minmu = 0.5,
  begin0based = TRUE)
}
\arguments{
\item{object}{a DESeqDataSet, on which one of the following functions has already been called:
\code{\link[DESeq2:nbinomWaldTest]{nbinomWaldTest}}\cr
\code{\link[DESeq2:nbinomLRT]{nbinomLRT}} is NOT supported in this version}

\item{annotationFile}{sliding window annotation file, can be plain either text or .gz file,
the file MUST be <TAB> separated. See Details for a description of the columns and column names expected in this file.}

\item{contrast}{this argument specifies what comparison to extract from the \code{object} to build a results table.}

\item{name}{the name of the individual effect (coefficient) for building a results table.
\code{name} argument is ignored if \code{contrast} is specified}

\item{listValues}{check \code{\link[DESeq2:results]{results}} for details of this parameter}

\item{cooksCutoff}{theshold on Cook's distance}

\item{test}{this is automatically detected internally if not provided.}

\item{addMLE}{if \code{betaPrior=TRUE} was used}

\item{tidy}{whether to output the results table with rownames as a first column 'row'.
The table will also be coerced to \code{data.frame}}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel
execution using \code{BiocParallel}, see next argument \code{BPPARAM}}

\item{BPPARAM}{an optional parameter object passed internally
to \code{\link{bplapply}} when \code{parallel=TRUE}.
If not specified, the parameters last registered with
\code{\link{register}} will be used.}

\item{minmu}{lower bound on the estimated count (used when calculating contrasts)}

\item{begin0based}{TRUE (default) or FALSE. If TRUE, then the start positions in \code{annotationFile} are  considered to be 0-based}
}
\value{
data.frame
}
\description{
This is a modified version of the
     \code{\link[DESeq2:results]{results}} function.

This function uses chromosomal positions given in the \code{annotationFile}
to identify overlapping windows in \code{dds} object. For each window,
the number of overlapping windows are counted, and the p-value is
 adjusted for FWER using bonferroni correction.

For further details, please refer documentation for
 \code{\link[DESeq2:results]{results}} function in DESeq2 package
}
\details{
\code{annotationFile} is expected to have the following columns:
\itemize{
  \item \code{chromosome}: chromosome name
  \item \code{unique_id}: unique id of the window, \code{rownames(object)} must match this column
  \item \code{begin}: window start co-ordinate, see parameter \code{begin0based}
  \item \code{end}: window end co-ordinate
  \item \code{strand}: strand
  \item \code{gene_id}: gene id
  \item \code{gene_name}: gene name
  \item \code{gene_type}: gene type annotation
  \item \code{gene_region}: gene region
  \item \code{Nr_of_region}: number of the current region
  \item \code{Total_nr_of_region}: total number of regions
  \item \code{window_number}: window number
}

The output data.frame from this function will have all the columns listed
above, along with the DESeq2 \code{\link[DESeq2:results]{results}}
columns:\cr
\code{baseMean}, \code{log2FoldChange}, \code{lfcSE}, \code{stat},
\code{pvalue}, \code{padj}. Please consult DESeq2 vignettes for an
 explanation of these columns.
}
