% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLBP_K562_w50s20_chr1_to_10_filtered-data.R
\docType{data}
\name{SLBP_K562_w50s20_chr1_to_10_filtered}
\alias{SLBP_K562_w50s20_chr1_to_10_filtered}
\title{ENCODE eCLIP data for SLBP in K562, filtered, chromosomes 1 to 10}
\format{An object of class \code{"DESeq"};}
\usage{
data(SLBP_K562_w50s20_chr1_to_10_filtered)
}
\description{
This is ENCODE eCLIP data which was quantified by htseq-clip
in sliding-windows of max. length 50nt, the step size was 20.
This is not ideal data for DEWSeq since it is lacking replicates,
however was small enough for the inclusion of the package.
}
\examples{
library(DEWSeq)
data(SLBP_K562_w50s20_chr1_to_10_filtered)
SLBP_K562_w50s20_chr1_to_10_filtered
}
\keyword{datasets}
