% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slbpDds.R
\docType{data}
\name{slbpDds}
\alias{slbpDds}
\title{ENCODE eCLIP data SLBP in K562}
\format{An object of class \code{"DESeq"};}
\usage{
data(slbpDds)
}
\description{
This is a DESeq dataset object for ENCODE eCLIP data: SLBP in K562 cell lines
This is used as an example dataset for a runnable example. This dataset is the output
from running the example code for the function \code{\link{DESeqDataSetFromSlidingWindows}}
}
\examples{

data(slbpDds)
slbpDds
}
\keyword{datasets}
