% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slbpRegions.R
\docType{data}
\name{slbpRegions}
\alias{slbpRegions}
\title{ENCODE eCLIP data SLBP in K562}
\format{data.frame;}
\usage{
data(slbpRegions)
}
\description{
This is a DESeq results object for ENCODE eCLIP data: SLBP in K562 cell lines
This is used as an example dataset for a runnable example. This dataset is the output
from running the example code for the function \code{\link{extractRegions}}
}
\examples{

data(slbpRegions)
head(slbpRegions)
}
\keyword{datasets}
