% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slbpVst.R
\docType{data}
\name{slbpVst}
\alias{slbpVst}
\title{ENCODE eCLIP data SLBP in K562}
\format{matrix;}
\usage{
data(slbpVst)
}
\description{
This is a DESeq normalized sliding window count matrix ENCODE eCLIP data: SLBP in K562 cell lines
This is used as an example dataset for a runnable example. This dataset is the output
from running the example code for the function \code{\link[DESeq2:vst]{vst}}
}
\examples{

data(slbpVst)
head(slbpVst)
}
\keyword{datasets}
