% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toBED}
\alias{toBED}
\title{windows/regions to BED}
\usage{
toBED(windowRes, fileName, padjCol = "padj", padjThresh = 0.05,
  log2FoldChangeCol = "log2FoldChange", log2FoldChangeThresh = 1,
  trackName = "sliding windows", description = "sliding windows")
}
\arguments{
\item{windowRes}{output data.frame from \code{\link{createRegions}}}

\item{padjCol}{name of the adjusted pvalue column (default: padj)}

\item{padjThresh}{threshold for p-adjusted value (default: 0.05)}

\item{log2FoldChangeCol}{name of the log2foldchange column (default: log2FoldChange)}

\item{log2FoldChangeThresh}{threshold for log2foldchange value (default:1)}

\item{trackName}{name of this track, for visualization}

\item{description}{description of this track, for visualization}
}
\description{
given output of \code{\link{createRegions}} and significance thresholds, extract significant windows, 
regions (merged neighboring significant windows) and create a BED file for visualization
}
