% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topWindow.R
\name{topWindowStats}
\alias{topWindowStats}
\title{stats for the top windows in each region}
\usage{
topWindowStats(windowRes, padjCol = "padj", padjThresh = 0.05,
  log2FoldChangeCol = "log2FoldChange", log2FoldChangeThresh = 1,
  begin0based = TRUE, normalizedCounts, treatmentCols,
  treatmentName = "treatment", controlName = "control", op = "max")
}
\arguments{
\item{windowRes}{output data.frame from \code{\link{results_DEWSeq}}}

\item{padjCol}{name of the adjusted pvalue column (default: padj)}

\item{padjThresh}{threshold for p-adjusted value (default: 0.05)}

\item{log2FoldChangeCol}{name of the log2foldchange column (default: log2FoldChange)}

\item{log2FoldChangeThresh}{threshold for log2foldchange value (default:1)}

\item{begin0based}{TRUE (default) or FALSE. If TRUE, then the start positions in \code{windowRes} is  considered to be 0-based}

\item{normalizedCounts}{data.frame with normalized read counts per window. \code{rownames(normalizedCounts)} and \code{unique_id} column from \code{windoeRes} must match}

\item{treatmentCols}{column names in \code{normalizedCounts} for treatment/case samples. The remaining columns in the data.frame will be considered control samples}

\item{treatmentName}{name to append to mean treatment  (default: treatment)}

\item{controlName}{name for}
}
\description{
given window resutls and normalized counts, combine significant overlapping windows into regions and
for each region, pick two candidate winodws:\cr (i) with highest log2FoldChange and\cr (ii) with highest normalized mean in
treatment samples.\cr Return a data.frame with region information and stats, and for the selected windows stats:\cr
normalized mean expression in treatment and control samples and\cr individual expression in replicates
}
