% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{.readAnnotation}
\alias{.readAnnotation}
\title{read annotation data}
\usage{
.readAnnotation(fname, uniqIds = NULL)
}
\arguments{
\item{fname}{file name/path}

\item{uniqIds}{filter stable and keep annotation for these unique ids}
}
\description{
read annotation data for windows
The file MUST be tab separated and MUST have the following columns:
chromosome: chromosome name
unique_id: unique id of the window
begin: window start co-ordinate
end: window end co-ordinate
strand: strand
gene_id: gene id
gene_name: gene name
gene_type: gene type annotation
gene_region: gene region
Nr_of_region: number of the current region
Total_nr_of_region: total number of regions
window_number: window number
}
