% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractRegions}
\alias{extractRegions}
\title{extract significant regions}
\usage{
extractRegions(windowRes, padjCol = "padj", padjThresh = 0.05,
  log2FoldChangeCol = "log2FoldChange", log2FoldChangeThresh = 1,
  begin0based = TRUE)
}
\arguments{
\item{windowRes}{output data.frame from \code{\link{createRegions}}}

\item{padjCol}{name of the adjusted pvalue column (default: padj)}

\item{padjThresh}{threshold for p-adjusted value (default: 0.05)}

\item{log2FoldChangeCol}{name of the log2foldchange column (default: log2FoldChange)}

\item{log2FoldChangeThresh}{threshold for log2foldchange value (default:1)}

\item{begin0based}{if TRUE, window start positions are 0-based}
}
\value{
data.frame
}
\description{
extract significant windows from output of \code{\link{createRegions}} using \code{regionStartId} column,
merge these significant windows to regions and create the following columns for each significant region: \cr
  \code{padj_min}: min. padj value in the region \cr
  \code{padj_max}: max. padj value in the region \cr
  \code{padj_avg}: avg. padj value in the region \cr
  \code{log2FoldChange_min}: min. log 2 fold change in the region \cr
  \code{log2FoldChange_max}: max. log 2 fold change in the region \cr
  \code{log2FoldChange_avg}: avg. log 2 fold change in the region \cr
}
