% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeWindows.R
\name{mergeWindows}
\alias{mergeWindows}
\title{merge windows}
\usage{
mergeWindows(annRes, minDist = 0, padjWindow = "bonferroni",
  padjMethod = "BH", ncores = 5)
}
\arguments{
\item{annRes}{annotated DESEq2 table}

\item{minDist}{minimum distance between the sliding windows
default 0 to account for bed formatted window regions}

\item{padjWindow}{p-value adjustment method for window, can be one of 'local' or 'bonferroni'}

\item{padjMethod}{p-value adjustment method (multiple testing correction) for the whole table, can be any input for base function p.adjust}

\item{ncores}{number of cores to use}
}
\description{
merge annotated sliding windows
}
