% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeWindowsHelper_bonferroni.R
\name{mergeWindowsHelper_bonferroni}
\alias{mergeWindowsHelper_bonferroni}
\title{bonferroni correction}
\usage{
mergeWindowsHelper_bonferroni(annRes, minDist = 0)
}
\arguments{
\item{annRes}{annotated results (data.frame)}

\item{minDist}{minimum distance between non overlapping windows,
default: 0 to account for bed formatted window regions}
}
\description{
Helper function merge overlapping windows, correct p-values for these overlapping windows using bonferroni FWER
}
\keyword{internal}
