% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results_DEWSeq}
\alias{results_DEWSeq}
\title{extract DEWseq results}
\usage{
results_DEWSeq(object, contrast, name, listValues = c(1, -1),
  cooksCutoff, test, addMLE = FALSE, tidy = FALSE, parallel = FALSE,
  BPPARAM = bpparam(), minmu = 0.5)
}
\arguments{
\item{object}{a DESeqDataSet, on which one of the following functions has already been called:
\code{\link{DESeq}}, \code{\link{nbinomWaldTest}}
\code{\link{nbinomLRT}} is NOT supported in this version}

\item{contrast}{this argument specifies what comparison to extract from the \code{object} to build a results table.}

\item{name}{the name of the individual effect (coefficient) for building a results table.
\code{name} argument is ignored if \code{contrast} is specified}

\item{cooksCutoff}{theshold on Cook's distance}

\item{test}{this is automatically detected internally if not provided.}

\item{addMLE}{if \code{betaPrior=TRUE} was used}

\item{tidy}{whether to output the results table with rownames as a first column 'row'.
The table will also be coerced to \code{data.frame}}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel
execution using \code{BiocParallel}, see next argument \code{BPPARAM}}

\item{BPPARAM}{an optional parameter object passed internally
to \code{\link{bplapply}} when \code{parallel=TRUE}.
If not specified, the parameters last registered with
\code{\link{register}} will be used.}

\item{minmu}{lower bound on the estimated count (used when calculating contrasts)}
}
\description{
This is a hack of the DESeq2 results function
credits to the authors
This documentation is based on DESeq2 results function documentation
Extract results from a DESeq analysis object
\code{results_DEWSeq} extracts a result table from a DESeq analysis giving base means across samples,
log2 fold changes, standard errors, test statistics, p-values and adjusted p-values;
For further details, please refer documentation for \code{results} function in DESeq2 package
}
