% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBED.R
\name{toBED}
\alias{toBED}
\title{windows/regions to BED}
\usage{
toBED(windowRes, regionRes, fileName, padjCol = "padj",
  padjThresh = 0.05, log2FoldChangeCol = "log2FoldChange",
  log2FoldChangeThresh = 1, trackName = "sliding windows",
  description = "sliding windows")
}
\arguments{
\item{windowRes}{output data.frame from \code{\link{resultsDEWSeq}}}

\item{regionRes}{output data.frame from \code{\link{extractRegions}}}

\item{fileName}{filename to save BED output}

\item{padjCol}{name of the adjusted pvalue column (default: padj)}

\item{padjThresh}{threshold for p-adjusted value (default: 0.05)}

\item{log2FoldChangeCol}{name of the log2foldchange column (default: log2FoldChange)}

\item{log2FoldChangeThresh}{threshold for log2foldchange value (default:1)}

\item{trackName}{name of this track, for visualization}

\item{description}{description of this track, for visualization}
}
\value{
write to file
}
\description{
given output of \code{\link{extractRegions}}, \code{\link{resultsDEWSeq}} and significance thresholds,
extract significant windows, create regions by merging adjacent significant windows.
Finally, write the output as a BED file for visualization.
}
\examples{
# need specific examples
\dontrun{
'toBED(windowRes=windowRes,regionRes=regionRes,fileName="enrichedWindowsRegions.bed")'
}

}
